+++ solidity/test/utils/PragmaTestUtils.t.sol
@@ -10,11 +10,10 @@ import {DataFeedType} from "../../src/interfaces/IPragma.sol";
 
 abstract contract PragmaTestUtils is Test, RandTestUtils {
     uint16 constant SOURCE_EMITTER_CHAIN_ID = 0x1;
-    bytes32 constant SOURCE_EMITTER_ADDRESS =
-        0x03dA250675D8c2BB7cef7E1b7FDFe17aA4D5752Ed82A9333e4F9a12b22E521aa;
+    bytes32 constant SOURCE_EMITTER_ADDRESS = 0x03dA250675D8c2BB7cef7E1b7FDFe17aA4D5752Ed82A9333e4F9a12b22E521aa;
 
-    uint constant SINGLE_UPDATE_FEE_IN_WEI = 1;
-    uint constant VALID_TIME_PERIOD_IN_SECONDS = 60;
+    uint256 constant SINGLE_UPDATE_FEE_IN_WEI = 1;
+    uint256 constant VALID_TIME_PERIOD_IN_SECONDS = 60;
 
     function setUpPragma(address hyperlane) public returns (address) {
         uint16[] memory emitterChainIds = new uint16[](1);
@@ -24,17 +23,13 @@ abstract contract PragmaTestUtils is Test, RandTestUtils {
         emitterAddresses[0] = SOURCE_EMITTER_ADDRESS;
 
         Pragma pragma_ = new Pragma(
-            hyperlane,
-            emitterChainIds,
-            emitterAddresses,
-            VALID_TIME_PERIOD_IN_SECONDS,
-            SINGLE_UPDATE_FEE_IN_WEI
+            hyperlane, emitterChainIds, emitterAddresses, VALID_TIME_PERIOD_IN_SECONDS, SINGLE_UPDATE_FEE_IN_WEI
         );
 
         return address(pragma_);
     }
 
-    function singleUpdateFeeInWei() public pure returns (uint) {
+    function singleUpdateFeeInWei() public pure returns (uint256) {
         return SINGLE_UPDATE_FEE_IN_WEI;
     }
 
@@ -56,12 +51,14 @@ abstract contract PragmaTestUtils is Test, RandTestUtils {
         bool brokenSignature;
     }
 
-    function encodeDataFeedMessages(
-        DataFeedMessage[] memory dataFeedMessages
-    ) internal pure returns (bytes[] memory encodedDataFeedMessages) {
+    function encodeDataFeedMessages(DataFeedMessage[] memory dataFeedMessages)
+        internal
+        pure
+        returns (bytes[] memory encodedDataFeedMessages)
+    {
         encodedDataFeedMessages = new bytes[](dataFeedMessages.length);
 
-        for (uint i = 0; i < dataFeedMessages.length; i++) {
+        for (uint256 i = 0; i < dataFeedMessages.length; i++) {
             encodedDataFeedMessages[i] = abi.encodePacked(
                 uint8(DataFeedType.SpotMedian),
                 dataFeedMessages[i].dataId,
@@ -77,33 +74,21 @@ abstract contract PragmaTestUtils is Test, RandTestUtils {
         DataFeedMessage[] memory dataFeedMessages,
         MerkleUpdateConfig memory config
     ) internal returns (bytes memory hyMerkleUpdateData) {
-        bytes[] memory encodedDataFeedMessages = encodeDataFeedMessages(
-            dataFeedMessages
-        );
+        bytes[] memory encodedDataFeedMessages = encodeDataFeedMessages(dataFeedMessages);
 
-        (bytes20 rootDigest, bytes[] memory proofs) = MerkleTree
-            .constructProofs(encodedDataFeedMessages, config.depth);
+        (bytes20 rootDigest, bytes[] memory proofs) = MerkleTree.constructProofs(encodedDataFeedMessages, config.depth);
 
-        bytes memory hyperlanePayload = abi.encodePacked(
-            rootDigest
-        );
+        bytes memory hyperlanePayload = abi.encodePacked(rootDigest);
 
         bytes memory hyperlaneValidatorSignature = generateValidatorSignature(
-            0,
-            config.source_chain_id,
-            config.source_emitter_address,
-            0,
-            hyperlanePayload,
-            config.numSigners
+            0, config.source_chain_id, config.source_emitter_address, 0, hyperlanePayload, config.numSigners
         );
 
         if (config.brokenVaa) {
-            uint mutPos = getRandUint() % hyperlaneValidatorSignature.length;
+            uint256 mutPos = getRandUint() % hyperlaneValidatorSignature.length;
 
             // mutate the random position by 1 bit
-            hyperlaneValidatorSignature[mutPos] = bytes1(
-                uint8(hyperlaneValidatorSignature[mutPos]) ^ 1
-            );
+            hyperlaneValidatorSignature[mutPos] = bytes1(uint8(hyperlaneValidatorSignature[mutPos]) ^ 1);
         }
 
         hyMerkleUpdateData = abi.encodePacked(
@@ -115,12 +100,9 @@ abstract contract PragmaTestUtils is Test, RandTestUtils {
             uint8(dataFeedMessages.length)
         );
 
-        for (uint i = 0; i < dataFeedMessages.length; i++) {
+        for (uint256 i = 0; i < dataFeedMessages.length; i++) {
             hyMerkleUpdateData = abi.encodePacked(
-                hyMerkleUpdateData,
-                uint16(encodedDataFeedMessages[i].length),
-                encodedDataFeedMessages[i],
-                proofs[i]
+                hyMerkleUpdateData, uint16(encodedDataFeedMessages[i].length), encodedDataFeedMessages[i], proofs[i]
             );
         }
     }
